Remove unnecessary .DS_Store files and implement storage and data_services packages; add news_fetcher service for fetching articles from News API
