+++ backend/microservices/data_services/__init__.py
@@ -1,0 +1,8 @@
+# data_services package
+
+"""
+Data Services Package
+
+This package contains modules for fetching and processing data from external sources.
+It provides services for retrieving news articles and other content from APIs.
+"""
