--- backend/microservices/processing_service.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python3
-"""
-processing_service.py - Microservice for Data Processing
-Processes raw data (cleaning, deduplication) and stores it.
-"""
-
-def run_processing():
-    # TODO: Implement processing logic
-    print("Processing service is running...")
-
-if __name__ == '__main__':
-    run_processing()
