+++ backend/microservices/storage/__init__.py
@@ -1,0 +1,2 @@
+# Make the storage directory a proper Python package
+# This allows importing modules from this directory
